+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 - Removed the ability to create Brotli files and removed the setting (use gzip instead).
 
 ### Fixed
+- Fixed a bug in which the `cacheDuration` config setting was not being applied when the value was not an integer ([#536](https://github.com/putyourlightson/craft-blitz/issues/536)).
+- Fixed a bug in which the `__home__` URI was not responding with a 404 error when it should have ([#538](https://github.com/putyourlightson/craft-blitz/issues/538)).
 - Fixed a bug in which eager-loading of auto-injected elements was not being tracked on cached pages.
 
 ### Deprecated
@@ -28,7 +30,7 @@
 
 ## 4.4.7 - 2023-07-17
 ### Changed
-- Hardened checks against null responses to help avoid errors ([#531](https://github.com/putyourlightson/craft-blitz/issues/531)).
+- Hardened checks against null responses to help avoid errors ([#519](https://github.com/putyourlightson/craft-blitz/issues/519)).
 
 ### Fixed
 - Fixed a bug in which refreshing the cache could fail when using the Redis queue driver ([#522](https://github.com/putyourlightson/craft-blitz/issues/522)).